feat: add confirmation dialog before deleting products and transactions
Add a confirmation dialog to prevent accidental deletion of products and transactions. This ensures users are aware of the action they are about to perform, reducing the risk of unintentional data loss.

+++ app/admin-products/page.tsx
@@ -79,8 +79,11 @@ const AdminProducts = () => {
   }
 
   const handleDeleteProduct = async (id: string) => {
-    await deleteDoc(doc(db, 'products', id))
-    fetchProducts()
+    const isConfirmed = window.confirm('Yakin mau dihapus nih?')
+    if (isConfirmed) {
+      await deleteDoc(doc(db, 'products', id))
+      fetchProducts()
+    }
   }
 
   const handleSubmit = async () => {
